fix(preview): 保持热更新期间的图片预览上下文稳定
共享 Context 和 hooks 移出 Provider 的 Fast Refresh 边界，消费者统一使用稳定模块；可选 API 缺失时复用稳定空实现。更新预览生命周期文档并通过全量前端回归。
